Se modifica la estructura inicial del proyecto y se copia todo el contenido de la carpeta laravel a la raíz del proyecto. Como desarrollo final se crean interfaces de usuario relacionado a Productos y Transacciones (Ventas), se crean los respectivos Modelos, Vistas y Controladores de Recursos.

+++ database/migrations/2023_01_11_001148_create_products_table.php
@@ -17,10 +17,10 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('nombre');
             $table->string('referencia');
-            $table->integer('precio');
-            $table->integer('peso');
+            $table->unsignedBigInteger('precio');
+            $table->unsignedBigInteger('peso');
             $table->string('categoria');
-            $table->integer('stock');
+            $table->unsignedBigInteger('stock');
             $table->timestamps();
         });
     }
